train/requirements.txt
tensorflow-cpu
pandas
sklearn
nltk
gensim
matplotlib